pgwire: split the authentication code in its own files
This patch splits the pgwire authentication in its own files
and adds missing explanatory comments.

Release note: None
